Improved README. Minor Refactoring.
